+++ Manager/BaseEntityWithImageManager.php
@@ -321,14 +321,18 @@ class BaseEntityWithImageManager extends BaseEntityWithFileManager
             }
 
             foreach ($this->imageFormatChoices[$propertyName] as $format) {
-
                 $oldDestPath = $this->getFileAbsolutePath($entity, $propertyName, $format);
-                if(is_file($oldDestPath)) {
+
+                $oldDestPath = $this->pathRemoveDotSegment($oldDestPath);
+                $sourceFilepath = $this->pathRemoveDotSegment($sourceFilepath);
+
+                if(is_file($oldDestPath) && $oldDestPath != $sourceFilepath) {
                     unlink($oldDestPath);
                 }
 
                 $absoluteDestFilepath = $this->getFileAbsolutePath($entity, $propertyName, $format);
                 $absoluteDestDir = substr($absoluteDestFilepath, 0, strrpos($absoluteDestFilepath, '/'));
+
                 if(!is_dir($absoluteDestDir)){
                     mkdir($absoluteDestDir, 0777, true);
                 }
@@ -347,5 +351,4 @@ class BaseEntityWithImageManager extends BaseEntityWithFileManager
 
         return null;
     }
-
 }
